fix: uniffi-related code breaking test compilation
